fixed issue with loading cart where lat/long is null
* also added error snackbar when something goes wrong there

+++ admin/src/main/java/org/dicadeveloper/weplantaforest/admin/dev/DatabasePopulator.java
@@ -233,8 +233,8 @@ public class DatabasePopulator {
 
             Tree treeDto = new Tree();
             treeDto.setAmount(i % 20 + 1);
-            treeDto.setLatitude(i);
-            treeDto.setLongitude(i);
+            treeDto.setLatitude((float) i);
+            treeDto.setLongitude((float) i);
             treeDto.setTreeType(articleToPlant.getTreeType());
             treeDto.setPlantedOn(timeOfPlant - i * 100000000L);
             treeDto.setOwner(userRepository.findByName(DEFAULT_USERS.get(i % 15)));
